+++ src/components/products/drag-dropzone.svelte
