+++ src/components/products/image-preview-card.svelte
