+++ src/components/products/preview-card.svelte
