+++ src/constants/products/blank-product.js
