+++ src/containers/products/item-content.svelte
