+++ src/routes/products/[id([0-9]+)]/edit.svelte
@@ -19,14 +19,14 @@
 <script>
   import { prefetch, goto } from '@sapper/app';
   import Layout from '@/layouts/default.svelte';
-  import Form from '@/containers/product/edit-form.svelte';
-  import PreviewCard from '@/components/product/preview-card.svelte';
+  import Form from '@/containers/products/edit-form.svelte';
+  import PreviewCard from '@/components/products/preview-card.svelte';
   import Button from 'ui/button.svelte';
   import Modal from 'ui/modal.svelte';
   import Dialog from 'ui/dialog.svelte';
   import * as api from '@/utils/api.js';
   import arraysEqual from '@/utils/arrays-equal.js';
-  import { getBlankVariety } from '@/constants/product/blank-product.js';
+  import { getBlankVariety } from '@/constants/products/blank-product.js';
 
   export let unmodifiedProduct;
   let product = Object.assign({}, unmodifiedProduct);
