+++ src/routes/products/new.svelte
@@ -18,13 +18,13 @@
   import { onMount } from 'svelte';
   import { prefetch, goto } from '@sapper/app';
   import Layout from '@/layouts/default.svelte';
-  import Form from '@/containers/product/edit-form.svelte';
-  import PreviewCard from '@/components/product/preview-card.svelte';
+  import Form from '@/containers/products/edit-form.svelte';
+  import PreviewCard from '@/components/products/preview-card.svelte';
   import Button from 'ui/button.svelte';
   import Modal from 'ui/modal.svelte';
   import Dialog from 'ui/dialog.svelte';
   import * as api from '@/utils/api.js';
-  import { getBlankProduct, getBlankVariety } from '@/constants/product/blank-product.js';
+  import { getBlankProduct, getBlankVariety } from '@/constants/products/blank-product.js';
 
   export let colors;
   export let sizes;
